+++ platform-sdk/platform-apps/tests/PlatformTestingTool/build.gradle.kts
@@ -22,6 +22,7 @@ testModuleInfo {
     requires("org.junit.jupiter.api")
     requires("org.mockito")
     requires("org.assertj.core")
+    requires("com.swirlds.merkledb.test.fixtures")
 }
 
 timingSensitiveModuleInfo {
