+++ platform-sdk/swirlds-platform-core/build.gradle.kts
@@ -39,6 +39,7 @@ testModuleInfo {
     requires("com.swirlds.state.api.test.fixtures")
     requires("com.swirlds.state.impl")
     requires("com.swirlds.state.impl.test.fixtures")
+    requires("com.swirlds.merkledb.test.fixtures")
     requires("org.hiero.base.crypto.test.fixtures")
     requires("org.hiero.base.utility.test.fixtures")
     requires("org.hiero.consensus.model.test.fixtures")
