+++ platform-sdk/swirlds-platform-core/src/testFixtures/java/module-info.java
@@ -23,6 +23,7 @@ open module com.swirlds.platform.core.test.fixtures {
     requires transitive org.junit.jupiter.api;
     requires com.swirlds.base.test.fixtures;
     requires com.swirlds.logging;
+    requires com.swirlds.merkledb.test.fixtures;
     requires com.swirlds.merkledb;
     requires com.swirlds.state.api.test.fixtures;
     requires org.hiero.base.crypto.test.fixtures;
