+++ platform-sdk/swirlds-state-impl/build.gradle.kts
@@ -9,8 +9,6 @@ testModuleInfo {
     requires("org.assertj.core")
     requires("com.swirlds.state.api.test.fixtures")
     requires("com.swirlds.state.impl.test.fixtures")
-    requires("com.swirlds.merkledb")
-    requires("com.swirlds.common.test.fixtures")
     requires("org.junit.jupiter.api")
     requires("org.mockito")
     requires("org.mockito.junit.jupiter")
