+++ platform-sdk/swirlds-virtualmap/build.gradle.kts
@@ -41,13 +41,13 @@ testModuleInfo {
 timingSensitiveModuleInfo {
     requires("com.swirlds.base")
     requires("com.swirlds.common")
-    requires("com.swirlds.merkledb")
     requires("com.swirlds.common.test.fixtures")
     requires("com.swirlds.config.api")
     requires("com.swirlds.config.extensions")
     requires("com.swirlds.config.extensions.test.fixtures")
     requires("com.swirlds.metrics.api")
     requires("com.swirlds.virtualmap.test.fixtures")
+    requires("com.swirlds.merkledb.test.fixtures")
     requires("org.hiero.base.crypto")
     requires("org.hiero.consensus.model")
     requires("org.hiero.base.concurrent")
